pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
feedparser = "*"
requests = "*"
beautifulsoup4 = "*"
lxml = "*"
python-dateutil = "^2.9.0.post0"
mysql-connector-python = "^9.6.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
